requirements.txt
absl-py==1.4.0
annotated-types==0.5.0
anyio==3.7.1
astunparse==1.6.3
av==10.0.0
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.6
cmake==3.26.4
contourpy==1.1.0
cycler==0.11.0
dnspython==2.4.0
email-validator==2.0.0.post2
exceptiongroup==1.1.2
fastapi==0.100.0
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.41.0
fsspec==2023.6.0
gast==0.4.0
gdown==4.7.1
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.0
h11==0.14.0
h5py==3.9.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
huggingface-hub==0.16.4
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.13.1
kiwisolver==1.4.4
libclang==16.0.6
lit==16.0.6
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opencv-python==4.8.0.74
opt-einsum==3.3.0
orjson==3.9.2
packaging==23.1
Pillow==10.0.0
protobuf==4.23.4
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.0.3
pydantic-extra-types==2.0.0
pydantic-settings==2.0.2
pydantic_core==2.3.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
retina-face==0.0.13
rsa==4.9
safetensors==0.3.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.27.0
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
tokenizers==0.13.3
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.65.0
transformers==4.31.0
triton==2.0.0
typing_extensions==4.7.1
ujson==5.8.0
urllib3==1.26.16
uvicorn==0.23.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
Werkzeug==2.3.6
wrapt==1.15.0
